[src] Split blocks into convolutional.py and recurrent.py
